Complete ex. 4.21

+++ osa4/controllers/blogs.js
@@ -41,6 +41,19 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+    const token = request.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    const userID = decodedToken.id
+    const toBeDeleted = await Blog.findById(request.params.id)
+
+    if (!token || !userID) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    if (userID != toBeDeleted.user) {
+        return response.status(401).json({ error: 'not authorized to delete this blog' })
+    }
+    
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
 })
